requirements.txt
python-telegram-bot==20.7
slack-sdk==3.26.1
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pandas==2.1.4
python-dateutil==2.8.2
feedparser==6.0.10 